Created platform dependent and independent classes for better separation. Platform independent stuff is now in 'common', while the platform dependent stuff is in the corresponding platform directory (win, linux mac).
