+++ common/logging.cpp
