+++ common/logging.h
