+++ win/platformcontext.cpp
@@ -1,0 +1,160 @@
+/*
+
+    OpenPnp-Capture: a video capture subsystem.
+
+    Windows platform code
+
+    Created by Niels Moseley on 7/6/17.
+    Copyright © 2017 Niels Moseley. All rights reserved.
+
+    Platform/implementation specific structures
+    and typedefs.
+
+*/
+
+//#include <Mfidl.h>
+//#include <Mfapi.h>
+
+#include <vector>
+#include "platformcontext.h"
+#include "../common/logging.h"
+#include "scopedcomptr.h"
+#include "../common/stream.h"
+
+PlatformContext::PlatformContext() :
+    Context()
+{
+    HRESULT hr;
+    hr = CoInitializeEx(NULL, COINIT_MULTITHREADED);
+    if (hr != S_OK)
+    {
+        LOG(LOG_CRIT, "Context creation failed!\n");
+        //FIXME: ok, what now? is it acceptable to throw an
+        //       exception here?
+    }
+    else
+    {
+        LOG(LOG_DEBUG, "Context created\n");
+    }
+
+    enumerateDevices();
+}
+
+PlatformContext::~PlatformContext()
+{
+    CoUninitialize();
+}
+
+bool PlatformContext::enumerateDevices()
+{
+    ICreateDevEnum*		dev_enum = nullptr;
+	IEnumMoniker*		enum_moniker = nullptr;
+	IMoniker*			moniker = nullptr;
+	IPropertyBag*		pbag = nullptr;
+
+    LOG(LOG_DEBUG, "Enumerating devices\n");
+
+    m_devices.clear();
+
+	//create an enumerator for video input devices
+	HRESULT hr = CoCreateInstance(CLSID_SystemDeviceEnum, NULL,CLSCTX_INPROC_SERVER,IID_ICreateDevEnum,(void**) &dev_enum);
+	if ((hr != S_OK) || (dev_enum == nullptr))
+    {
+        LOG(LOG_CRIT, "Could not create ICreateDevEnum object\n");
+        return false;
+    }
+    else
+    {
+        LOG(LOG_DEBUG, "ICreateDevEnum created\n");
+    }
+
+    ScopedComPtr<ICreateDevEnum> devEnum(dev_enum);
+
+	hr = devEnum->CreateClassEnumerator(CLSID_VideoInputDeviceCategory,&enum_moniker,NULL);
+	if (hr == S_FALSE)
+    {
+        // no devices found!
+        LOG(LOG_INFO, "No devices found\n");
+        return true;
+    }
+    if (hr != S_OK)
+    {
+        LOG(LOG_CRIT, "Could not create class enumerator object\n");
+        return false;
+    }
+
+    ScopedComPtr<IEnumMoniker> enumMoniker(enum_moniker);
+
+	//get devices
+	uint32_t num_devices = 0;
+    VARIANT name;
+	while (enumMoniker->Next(1, &moniker,0) == S_OK)
+	{
+		//get properties
+		hr = moniker->BindToStorage(0, 0, IID_IPropertyBag, (void**) &pbag);
+        ScopedComPtr<IPropertyBag> propertyBag(pbag);
+
+		if (hr >= 0)
+		{            
+            deviceInfo info;
+			VariantInit(&name);
+
+			//get the description
+			hr = propertyBag->Read(L"Description", &name, 0);
+			if (hr < 0) hr = propertyBag->Read(L"FriendlyName", &name, 0);
+			if (hr >= 0)
+			{
+				BSTR BStringPtr = name.bstrVal;
+
+                if (BStringPtr)
+                {
+                    // copy wchar device name into info structure so we can reference the
+                    // device later
+                    info.m_filterName = std::wstring(BStringPtr);
+
+                    // convert wchar string to UTF-8 to pass to JAVA                    
+                    info.m_name = wcharPtrToString(BStringPtr);
+                }                
+            }
+            else
+            {
+                LOG(LOG_ERR, "Could not generate device name for device!\n");
+            }
+
+            hr = propertyBag->Read(L"DevicePath", &name, 0);
+			if (hr >= 0)
+			{
+        		BSTR BStringPtr = name.bstrVal;            
+                if (BStringPtr)
+                {
+                    info.m_devicePath = std::wstring(BStringPtr);
+                    LOG(LOG_INFO, "     -> PATH %s\n", wcharPtrToString(BStringPtr).c_str());                    
+                } 
+            }
+
+            moniker->AddRef();
+            info.m_moniker = moniker;
+            m_devices.push_back(info);
+            LOG(LOG_INFO, "ID %d -> %s\n", num_devices, info.m_name.c_str());            
+        }
+
+        num_devices++;
+    }
+    return true;
+}
+
+std::string PlatformContext::wstringToString(const std::wstring &wstr)
+{
+    return wcharPtrToString(wstr.c_str());
+}
+
+std::string PlatformContext::wcharPtrToString(const wchar_t *sstr)
+{
+    std::vector<char> buffer;
+    int32_t chars = WideCharToMultiByte(CP_UTF8, 0, sstr, -1, nullptr, 0, nullptr, nullptr);    
+    if (chars == 0) return std::string("");
+
+    buffer.resize(chars);
+    WideCharToMultiByte(CP_UTF8, 0, sstr, -1, &buffer[0], chars, nullptr, nullptr);
+    return std::string(&buffer[0]);
+}
