+++ win/platformstream.h
@@ -11,8 +11,8 @@
 
 */
 
-#ifndef stream_h
-#define stream_h
+#ifndef win32platform_stream_h
+#define win32platform_stream_h
 
 #include <windows.h>
 #include <dshow.h>
@@ -22,12 +22,14 @@
 #include <stdint.h>
 #include <vector>
 #include <mutex>
-#include "logging.h"
+#include "../common/logging.h"
+#include "../common/stream.h"
 #include "samplegrabber.h"
 
-class Context;      // pre-declaration
-struct deviceInfo;  // pre-declaration
-class Stream;       // pre-declaration
+
+class Context;          // pre-declaration
+struct deviceInfo;      // pre-declaration
+class PlatformStream;   // pre-declaration
 
 /** A class to handle callbacks from the video subsystem,
     A call is made for every frame.
@@ -39,7 +41,7 @@ class Stream;       // pre-declaration
 class StreamCallbackHandler : public ISampleGrabberCB
 {
 public:
-    StreamCallbackHandler(Stream *stream) : m_stream(stream)
+    StreamCallbackHandler(PlatformStream *stream) : m_stream(stream)
     {
         m_callbackCounter = 0;
     }
@@ -94,28 +96,27 @@ public:
 
 
 private:
-    std::mutex  m_bufferMutex;      ///< mutex to protect frame buffer access
-    Stream*     m_stream;
-    uint32_t    m_callbackCounter;
+    PlatformStream* m_stream;
+    uint32_t        m_callbackCounter;
 };
 
 
 /** The stream class handles the capturing of a single device */
-class Stream
+class PlatformStream : public Stream
 {
 friend StreamCallbackHandler;
 
 public:
-    Stream();
-    ~Stream();
+    PlatformStream();
+    ~PlatformStream();
 
     /** Open a capture stream to a device and request a specific (internal) stream format. 
         When succesfully opened, capturing starts immediately.
     */
-    bool open(Context *owner, deviceInfo *device, uint32_t width, uint32_t height, uint32_t fourCC);
+    virtual bool open(Context *owner, deviceInfo *device, uint32_t width, uint32_t height, uint32_t fourCC) override;
 
     /** Close a capture stream */
-    void close();
+    virtual void close() override;
 
     /** Returns true if a new frame is available for reading using 'captureFrame'. 
         The internal new frame flag is reset by captureFrame.
@@ -127,33 +128,12 @@ public:
         must be supplied in RGBbufferBytes.
     */
     bool captureFrame(uint8_t *RGBbufferPtr, uint32_t RGBbufferBytes);
-    
-    /** Returns true if the stream is open and capturing */
-    bool isOpen() const
-    {
-        return m_isOpen;
-    }
 
     /** Return the FOURCC media type of the stream */
-    uint32_t getFOURCC();
-
-    /** Return the number of frames captured.
-        FIXME: protect by mutex 
-    */
-    uint32_t getFrameCount() const
-    {
-        return m_frames;
-    }
+    virtual uint32_t getFOURCC() override;
 
 protected:
-    void submitBuffer(uint8_t* ptr, size_t bytes);
     void dumpCameraProperties();
-
-    Context*    m_owner;                    ///< The context object associated with this stream
-
-    uint32_t    m_width;                    ///< The width of the frame in pixels
-    uint32_t    m_height;                   ///< The height of the frame in pixels
-    bool        m_isOpen;
 
     IFilterGraph2*  m_graph;
     IMediaControl*  m_control;
@@ -166,11 +146,6 @@ protected:
     /** Each time a new frame is available, the DirectShow subsystem
         will call the callback handler */
     StreamCallbackHandler *m_callbackHandler;
-
-    std::mutex  m_bufferMutex;              ///< mutex to protect m_frameBuffer and m_newFrame
-    bool        m_newFrame;                 ///< new frame buffer flag
-    std::vector<uint8_t> m_frameBuffer;     ///< raw frame buffer
-    uint32_t    m_frames;                   ///< number of frames captured
 
     VIDEOINFOHEADER m_videoInfo;            ///< video information of current captured stream
 };
