+++ win/tests/CMakeLists.txt
@@ -16,7 +16,7 @@ project (openpnp-capture-test)
 # add include directory 
 include_directories(../include ..)
 
-set (SOURCE main.cpp ../logging.cpp)
+set (SOURCE main.cpp ../../common/logging.cpp)
 
 add_executable(openpnp-capture-test ${SOURCE})
